fastapp/confsettings/config.py
from pydantic_settings import BaseSettings , SettingsConfigDict

class Settings(BaseSettings):

    MONGO_URI: str
    SECRET_KEY: str
    ALGORITHM: str
    ACCESS_TOKEN_EXPIRE_MINUTES: int    
    REFRESH_TOKEN_EXPIRE_DAYS: int

    model_config = SettingsConfigDict(
        env_file=".env", 
        env_file_encoding="utf-8",
        extra ="ignore"
    )

CONFIG = Settings()