fastapp/requirements.txt
a2wsgi==1.10.10
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
apache-airflow==3.0.6
apache-airflow-core==3.0.6
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.6.2
apache-airflow-providers-common-sql==1.28.0
apache-airflow-providers-smtp==2.2.1
apache-airflow-providers-standard==1.7.0
apache-airflow-task-sdk==1.0.6
APScheduler==3.11.0
argcomplete==3.6.2
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cadwyn==5.4.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==45.0.7
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
fastapi-cli==0.0.11
fastapi-cloud-cli==0.1.5
Flask==3.1.2
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
libcst==1.8.4
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
motor==3.7.1
msgspec==0.19.0
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
protobuf==6.32.0
psutil==7.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
PyYAML-ft==8.0.0
referencing==0.36.2
requests==2.32.5
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
sentry-sdk==2.37.1
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
starlette==0.47.3
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
typer==0.17.4
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
zipp==3.23.0
